refactor: API path 수정
